admin menu cateogry feature

+++ src/main/java/com/by/repository/UserRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.by.model.User;
 
-public interface UserRepository extends PagingAndSortingRepository<User, Long> {
+public interface UserRepository extends PagingAndSortingRepository<User, Integer> {
 	User findByName(String name);
 
-	Optional<User> findById(Long id);
+	Optional<User> findById(Integer id);
 }
