+++ src/main/java/com/by/service/impl/UserServiceImpl.java
@@ -1,21 +1,31 @@
 package com.by.service.impl;
 
-import com.by.form.UserQueryForm;
-import com.by.model.Menu;
-import com.by.model.User;
-import com.by.repository.UserRepository;
-import com.by.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.*;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 
 import javax.persistence.EntityManager;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
-import java.util.*;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.by.form.UserQueryForm;
+import com.by.model.Menu;
+import com.by.model.User;
+import com.by.repository.UserRepository;
+import com.by.service.UserService;
 
 @Service
 @Transactional
@@ -82,7 +92,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional(readOnly = true)
-    public Optional<User> findId(Long id) {
+    public Optional<User> findId(int id) {
         return repository.findById(id);
     }
 }
